refactor: 상품 관련 로직 분리
